requirements.txt
telethon>=1.36,<2.0
discord.py>=2.3,<3.0
fastapi>=0.104
uvicorn[standard]>=0.24
# Security dependencies
cryptography>=41.0.0  # For secure data handling
